I created a bin folder with resources for the javafx library to use when we compile.
I updated the main function to have the variables to hold scenes, and the ability to
intialize them with the init() method.

I also added the titleScreen class and TitleController class to the GUI folder. However
the title screen is not completely finished, and just displays a template.

My current java editor and compiler is giving me some troubles, and I can not run the code
I have placed. This may be because of a java error on my end. If it doesn't compile, let me know,
and I will try to see what needs to be changed.